fix: When choosing an algorithm suite, select the algorithm suite based on the configuration of the dh parameter.
Co-authored-by: pi_ixeL<xuzhewei2@huawei.com>



# message auto-generated for no-merge-commit merge:
!851 merge fix_ into main

fix: When choosing an algorithm suite, select the algorithm suite based on the configuration of the dh parameter.

Created-by: pi_ixeL
Commit-by: pi_ixeL
Merged-by: liwei3013
Description: fix: fix following issue
When choosing an algorithm suite, select the algorithm suite based on the configuration of the dh parameter.
In the TLS 1.3 scenario, if no certificate or PSK is configured, an alert will be sent.

See merge request: openHiTLS/openhitls!851